# Cache optimization