A small patch to enhance cache with Ansible.

Used playbook **huge-inventory.yml**:

Host count | Before | After
-----------|--------|------
100        | 3.1s   | 2.5s
200        | 6.6s   | 5.39s

Playbook content:
```yaml
- hosts: compute
  gather_facts: no
  tasks:
    - debug: msg={{inventory_hostname}}
      delegate_to: localhost
```