@jimi-c 2015-11-09
Merged, thanks!