@jimi-c 2015-11-09
For the record, using this on an inventory with 1000 hosts shaved off about 4s (43.5s vs. 47.5s on my system), so about 10% boost there.